0. _putchar

+++ 0x02-functions_nested_loops/0-putchar.c
@@ -10,7 +10,7 @@ int main(void)
 	int i = 0;
 	char strArr[] = "_putchar\n";
 
-	for(i = 0; i < sizeof(strArr); ++i)
+	for(i = 0; i < (int)sizeof(strArr); ++i)
 	{
 		_putchar(strArr[i];
 	}
